refactor: delete curation adapters and clean scheduling dead code
Remove the entire curation adapter stack (docs, youtube, ranking, validate,
config, types) and Google API rate limiter — all production-unused. Delete
dead scheduling wrappers and unused date helpers. Deduplicate validation enums.

Changes:
- Delete src/lib/curation/ entirely (6 files, ~1390 LOC)
- Delete src/lib/utils/google-api-rate-limiter.ts (499 LOC)
- Delete scheduling/validate.ts and scheduling/generate.ts (test-only wrappers)
- Remove 4 unused date helpers from scheduling/dates.ts
- Extract SKILL_LEVEL_ENUM and LEARNING_STYLE_ENUM to validation/shared.ts
- Remove unused learningPlanResourceSchema from validation/learningPlans.ts
- Delete 6 related test files and fixture
- Update schedule.ts to import from distribute.ts directly
